requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
python-multipart==0.0.9
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.3.0
pypdf==5.1.0
pymupdf==1.24.10
pytesseract==0.3.13
Pillow==10.4.0
pytest==8.3.3
httpx==0.27.2
